✅ Add specs for batch email notifications
This commit will add a few specs and also redo the logic for the batch
email notification send frequency.  We needed a migration to the users
model to track when the last email was sent to check if it's time to
send another one based on their batch_email_frequency setting.