🎁 Adds symlink between samvera and hyrax-webapp branding directories
The issue of branding, logo, and thumbnails not appearing has been solved before in other projects.
ref:
- https://github.com/scientist-softserv/atla-hyku/pull/187

This commit is required to resolve the issue with uploading images from the Collections' branding tab. Because the symlink wasn't established, a user would not be able to see the images they uploaded even though the path correctly exists.