🐛 Fix gallery and masonry catalog results styling
Previously, when the user looks at the catalog search results page and
sets the view to either gallery or masonry, the styling is really off.
Here this commit should fix that and also hide the bookmarks checkbox.
Hyrax dassie does not show it so we will mimic that until further notice
since the user can't even get to the /bookmarks route.

Ref:
  - https://github.com/scientist-softserv/hykuup_knapsack/issues/229