Change the structure of work show, add child link
Since the bootstrap upgrade the structure of ul > il > a doesn't seem to
work like before.  The <a> tag was only clickable when you hover over
the text.  This change will make the whole <li> clickable much like the
nav dropdown menu for the My Profile/Dashboard/Logout links on the top
right of the masthead.