propagate locale between group form tabs
This is a UX improvement. However, it also fixes a bug where some users
were getting an error message on these pages for the locale param being
an empty string (`?locale=`). Strangely, this only appeared to be
affecting users using Windows, but that may not be causation.

Originally implemented in PALNI/PALCI's Hyku:
- https://github.com/scientist-softserv/palni-palci/pull/1041/commits/6bb1f07ca5b7fcac760177a811decf0c87a48d29