🐛 Remove outdated blacklight_range_limit partial
For any applications that use the Blacklight Range Limit facet, we were
getting an error when the facets rendered with a proper date because we
had an override of an older version of a Blacklight Range Limit gem that
does not exist in the current (8.4.0) version we are using.  Also we
need to override Blacklight::Controller for the Cultural Repository
theme because it has the facets on the Homepage.  Without the override
we get a no route error.