✅ Fix Flaky featured_collection_list_spec
This commit will address a flaky spec in the FeaturedCollectionList
spec.  The collection titles are sequenced like "Collection Title 1",
"Collection Title 2", etc and one of the tests checks if the featured
collection list had the collection in title sorted order.  If the test
is ran after a lot of other collections have been created, we get a
situation where the titles are "Collection Title 9" and "Collection
Title 10".  The test asserts that it should be in the order of 9 then
10, however, the actual sorting is 10 then 9 because of the way Ruby
sorts the strings.

```rb
["9", "3", "10"].sort #=> ["10", "3", "9"]
```

The solution I propose is to override the sequenced collections so it'll
always be 1 and 2 and can't tick all the way up into the double digits.

I also the `:clean_repo` flag since I noticed that each run, there are
more and more collections that accumulate.