♻️ Introduce the `Hyku::Forms::Admin::Appearance`
In the HykuKnapsack we are running into an issue with a decorator that
was trying to alter the `#customization_params` method of the
`Hyrax::Forms::Admin::Appearance` however that method was being declared
in a decorator in Hyku.  The decorator in Knapsack was trying to alter
it prior to the method being declared and throwing a `NoMethodError`.
This refactor will make the `Hyku::Forms::Admin::Appearance` inherit
from the Hyrax version and override methods as needed thus declaring the
`#customization_params` method so the HykuKnapsack decorator can alter
it.