💄 Match catalog search results match to Hyrax
This commit will along the look of the catalog search results page to
look closer to the way Hyrax looks.  The problem stems from the
markdown feature.  For plain text it adds a `<p>` tag which has a margin
bottom.  I think the best thing to do is turn it off for default and any
tenant that is currently using it should turn it back on.